refactor: centralize server API models
